Move save stuff into crate::session::save.
